Refactor SOME icons
