+++ apps/web/src/components/icons/GitHubIcon.tsx
@@ -1,4 +1,4 @@
-export const GitHubIcon = () => (
+const GitHubIcon = () => (
   <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
     <rect x="0.648682" width="73.3513" height="73" rx="36.5" fill="#FAB759" />
     <path
@@ -9,3 +9,5 @@ export const GitHubIcon = () => (
     />
   </svg>
 )
+
+export default GitHubIcon
