+++ apps/web/src/components/icons/index.ts
@@ -1,5 +1,8 @@
 export { default as BedpressIcon } from "./BedpressIcon"
-export { default as GithubIcon } from "./GithubIcon"
+export { default as GitHubIcon } from "./GitHubIcon"
+export { default as SlackIcon } from "./SlackIcon"
+export { default as InstagramIcon } from "./InstagramIcon"
+export { default as FacebookIcon } from "./FacebookIcon"
 export { default as ItexIcon } from "./ItexIcon"
 export { default as LinkedinIcon } from "./LinkedinIcon"
 export { default as NextjsIcon } from "./NextjsIcon"
